make fetching of locations earlier in the code to prevent additional request to satellite
