+++ sharing/standard.go
@@ -5,11 +5,16 @@ package sharing
 
 import (
 	"context"
+	"encoding/json"
 	"net/http"
 	"net/url"
 	"strings"
 
 	"github.com/zeebo/errs"
+	"go.uber.org/zap"
+
+	"storj.io/uplink"
+	"storj.io/uplink/private/object"
 )
 
 func (handler *Handler) handleStandard(ctx context.Context, w http.ResponseWriter, r *http.Request) (err error) {
@@ -56,6 +61,50 @@ func (handler *Handler) handleStandard(ctx context.Context, w http.ResponseWrite
 	}
 
 	pr.access = access
+
+	q := r.URL.Query()
+
+	// flag used for get locations request.
+	locationsFlag := queryFlagLookup(q, "locations", false)
+	if locationsFlag {
+		ipBytes, err := object.GetObjectIPs(ctx, uplink.Config{}, pr.access, pr.bucket, pr.realKey)
+		if err != nil {
+			return WithAction(err, "get object IPs")
+		}
+
+		type location struct {
+			Latitude  float64
+			Longitude float64
+		}
+
+		// we explicitly don't want locations to be nil, so it doesn't
+		// render as null when we plop it into the output javascript.
+		locations := make([]location, 0, len(ipBytes))
+		if handler.mapper != nil {
+			for _, ip := range ipBytes {
+				info, err := handler.mapper.GetIPInfos(ctx, string(ip))
+				if err != nil {
+					handler.log.Error("failed to get IP info", zap.Error(err))
+					continue
+				}
+
+				locations = append(locations, location{
+					Latitude:  info.Location.Latitude,
+					Longitude: info.Location.Longitude,
+				})
+			}
+		}
+
+		w.Header().Set("Content-Type", "application/json")
+
+		err = json.NewEncoder(w).Encode(locations)
+		if err != nil {
+			handler.log.Error("failed to write json list locations response", zap.Error(err))
+		}
+
+		return nil
+	}
+
 	pr.visibleKey = pr.realKey
 	pr.title = pr.bucket
 	pr.root = breadcrumb{Prefix: pr.bucket, URL: "/s/" + serializedAccess + "/" + pr.bucket + "/"}
